requirements.txt
flask
pyYaml
pprint
elasticsearch==5.4.0